requirements.txt
customtkinter
pynput
packaging
pystray
Pillow
psutil
